Separation of compute and the solver:
We use template in the new function compute_for_solver which take in a template of class Method which is a solver in our case
then we call the static function compute iteration of the solver we used.
This method is usefull as we want to be able to use different kind of solver like Jacobi and Gauss seidel .
To do so we created a class Jacobi which contain the static method compute_iteration called by the template in compute_for_solver.