Debug  and vtk compatibility
In this commit we added a release version and moved all the print to debug mode.
This is important as we will later need to evaluate the performance of the methods and we don't
want costly cout call to slow us.

Additionnaly in this commit we added a function to variable that allow us to format our
result to vtk format , this allow us to vizualize our solution in Paraview later